requirements.txt
azure-cognitiveservices-speech==1.31.0
python-dotenv==1.0.0
keyboard==0.13.5
pyautogui==0.9.54
vosk==0.3.45
sounddevice==0.4.6
